Package Jax NNLS code
The NNLS problem is solved with a primal-dual interior point algorithm.  The qpax package was used as a basis for this code, but it has been stripped back for the specific case of NNLS.  This speeds up the code by a large amount on large problems (when Q is ~1000x1000 the speedup around 50x).